Add abstract pricer class, rename Conservative Pricer
You hopefully only ever want to price a line item, so that part of the name
was redundant. Also added an `Abstract` Pricer class that holds errors and the common signature
for pricers.
